* Move `Dockerfile` into project root. * Clean and rearrange. * Remove `source-extractor` methods.
